+++ packages/common/src/associations/types.ts
@@ -1,4 +1,58 @@
+import { HubEntityType } from "../core";
+import { IQuery } from "../search";
+
 /**
+ * associations are hierarchical in nature e.g.
+ * there is always a parent and a child in the
+ * relationship. This interface allows us to
+ * define these hierarchies on an entity-by-entity
+ * basis
+ */
+export interface IHubAssociationHierarchy {
+  children: HubEntityType[];
+  parents: HubEntityType[];
+}
+
+/**
+ * association rules stored on the parent entity
+ * that define what is "included" by the parent
+ *
+ * For now, the query will define an association
+ * group, and "included" simply means it has
+ * been shared with the association group. In
+ * the future, the query may contain additional
+ * conditions
+ */
+export interface IHubAssociationRules {
+  /** schema version for migration purposes */
+  schemaVersion: number;
+  /** query that defines what's "included" by a parent */
+  query: IQuery;
+}
+
+/**
+ * associations involve a 2-way agreement between
+ * parent and child. This interface allows us to
+ * keep track of an entity's association stats with
+ * another entity
+ */
+export interface IAssociationStats {
+  /** number of full associations */
+  associated: number;
+  /** number of outgoing association requests */
+  pending: number;
+  /** number of incoming association requests */
+  requesting: number;
+  /** number of entity's the child references = associated + pending */
+  referenced?: number;
+  /** number of entities included by the parent = associated + pending */
+  included?: number;
+}
+
+/**
+ * ** DEPRECATED: This will be removed in the next
+ * breaking version **
+ *
  * Definition of an Association
  * This will be persisted in the item's typekeywords
  * as `type|id`
@@ -15,7 +69,9 @@ export interface IAssociationInfo {
 }
 
 /**
+ * ** DEPRECATED: This will be removed in the next
+ * breaking version **
+ *
  * Association type
  */
 export type AssociationType = "initiative";
-// AS WE ADD MORE TYPES, UPDATE THE getItemTypeFromAssociationType FUNCTION