+++ packages/common/src/initiatives/_internal/InitiativeAssociationHierarchies.ts
@@ -1,0 +1,6 @@
+import { IHubAssociationHierarchy } from "../../associations/types";
+
+export const InitiativeAssociationHierarchies: IHubAssociationHierarchy = {
+  children: ["project"],
+  parents: [],
+};