+++ packages/common/src/projects/_internal/ProjectBusinessRules.ts
@@ -28,6 +28,7 @@ export const ProjectPermissions = [
   "hub:project:workspace:overview",
   "hub:project:workspace:dashboard",
   "hub:project:workspace:details",
+  "hub:project:workspace:initiatives",
   "hub:project:workspace:settings",
   "hub:project:workspace:collaborators",
   "hub:project:workspace:content",
@@ -101,6 +102,11 @@ export const ProjectPermissionPolicies: IPermissionPolicy[] = [
   },
   {
     permission: "hub:project:workspace:details",
+    dependencies: ["hub:project:workspace", "hub:project:edit"],
+  },
+  {
+    permission: "hub:project:workspace:initiatives",
+    availability: ["alpha"], // gate to just alpha for now
     dependencies: ["hub:project:workspace", "hub:project:edit"],
   },
   {